requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.6
Django==5.0.7
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-filter==24.2
django-js-asset==2.2.0
django-mailjet==0.3.1
django-six==1.0.5
dnspython==2.6.1
idna==3.7
mailjet-rest==1.3.4
pillow==10.4.0
py3dns==4.0.2
pycurl==7.45.3
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
